Handle file references which now contain a valid build phase
Previously file references that already existed but had a build phase set would be ignored. Unity 5 generates file references e.g AdSupport.framework that don't belong to any build phase. Change the way file references are added so that a change to a build phase will add that file to the required build phase. Required changing the build references to not allow duplicates